examples/inference/resnet50-python-tensorflow/requirements.txt
setuptools>=69.1.0
pillow==10.2.0
tensorflow==2.13.0
transformers==4.37.1
tritonclient[http]==2.42.0
